Update test_durations workflow (#2245)
### Ticket
https://github.com/tenstorrent/tt-forge-fe/issues/2108

### Problem description
For better CI split balance, we need to record the latest execution
durations for each test. To do so, we should update the details in the
.test_durations file.

### What's changed
New "Update Test Durations" workflow is introduced that updates test
durations based on the latest nightly and push workflow runs. It
processes the gathered XML test reports from the last successful nightly
and push workflows, calculates the test durations, and creates a pull
request to update the `.test_durations` file.

Workflow is NOT automated. It should be manually run when needed.

### Checklist
- [ ] New/Existing tests provide coverage for changes

Tested in Perf Benchmark workflow:
https://github.com/tenstorrent/tt-forge-fe/actions/runs/15557474048
Resulting PR: https://github.com/tenstorrent/tt-forge-fe/pull/2244