Add logical_not op in lower_to_mlir and add skip_model_analysis marker in paddleocr model (#2359)
1. The models ops logical_not test cases(i.e
`forge/test/models_ops/test_logicalnot.py`) was failing with
`RuntimeError: Found Unsupported operations while lowering from TTForge
to TTIR in forward graph` error but the logical_not was present in the
TTIR(i.e
[ttirlogical_not](https://docs.tenstorrent.com/tt-mlir/autogen/md/Dialect/TTIROp.html#ttirlogical_not-ttttirlogicalnotop))
so mapped it in lower to mlir and tested the logical_not models ops test
and the test passed. so removed xfail marker.

2. Added `skip_model_analysis` pytest marker for `paddleocr` model
because it contains two compile function and for these kind of cases
will not generate models ops tests.