Pytest crash/hang detector (#2397)
### Ticket
https://github.com/tenstorrent/tt-forge-fe/issues/2362

### Problem description
We experienced some pytest crashes (Abort, Out of memory, Segmentation
fault) as well as test hangs.
When these occur the problem is that all the tests running in that job
are treated as failed from analytics side and it is hard to track if
crashes are random or happen continually on particular tests.

### What's changed
Pytest is executed as a separate process so when it crashes it is
possible to detect it.
Also, before each test case is executed watchdog timer is set, so if
single test hangs longer then 15 min it is killed.
Crashed test summary is created and crash report uploaded as workflow
artifact.
After the crash is detected, test is removed from collection and pytest
is restarted. This is done to ensure that junit xml sent to analytics is
complete. There is an option "--continue-after-crash" which will remove
the test and continue with testing for test cases where analytics is
less important (eg. currently used in sweep tests).
There is also an option "--run-crashed-tests" which reruns crashed tests
one by one after the test suite is completed. It is currently unused,
but it might be activated in the future.

### Checklist
Test run with implemented hanging test which is killed by watchdog and
reported as crashed:
https://github.com/tenstorrent/tt-forge-fe/actions/runs/15875516814

- [ ] New/Existing tests provide coverage for changes