Added git-blame ignore file for formatting commits. (#2443)
Git has the ability to ignore provided commits in git-blame command
(https://git-scm.com/docs/git-config#Documentation/git-config.txt-blameignoreRevsFile).
This is useful for large formatting/refactoring commits that hides
initial changes.

This change adds a file that contains commits that are not important for
blaming. We can set this file as a default git-blame ignore file with:
$ git config blame.ignoreRevsFile .git-blame-ignore-revs