fix build file (#2445)
### Ticket
New TT-Forge docs have issue where it's referencing the wrong file name
for the build of the html docs

### Problem description
The wrong file is referenced, so the build page appears blank. 

### What's changed
Remove the build page, we already have the new build page under a
different title now.
This change will remove the blank page issue, and allow people to focus
on the new build page.

### Checklist
- [ ] New/Existing tests provide coverage for changes