[watchdog] show timout exceeded message
When running the tests with pytest capture enabled, the watchdog timeout
message is not being showed. This is because the watchdog will abort the
process, so pytest doesn't print out the captured logs.

If capture enabled, suspend the capture allowing us to print out the
message directly. Useful for CI, since i suspect there are watchdog
timeouts occuring in recent nightlies...