[autograd] reshape gradient outputs (#2614)
There are cases when the gradient outputs do not have the same rank as
their respective input nodes.

This can happen if we encounter `broadcast` ops in the forward graph.
The implementation of backward for `broadcast` is hacked inside the
`autograd.cpp` (outside of the regular op infrastructure) and doesn't
handle cases when `broadcast` needs to `unsqueeze`/`squeeze` its input.

This change checks for rank mismatch between the gradient output and its
respective input node, and inserts a reshape op. The reshape op will be
decomposed into `unsqeeze`/`squeeze` op (during post autograd
decompose).

Filed an issue #2612 to remove this workaround once we fix the backward
of `broadcast`.

The issue was discovered when implementing test infra for the ops. Those
unit tests will be added in upcoming PRs.