Fix ConvertEmulatedDtypes and remove sanity run before forge compilation in yolov5 (#2694)
The yolov5 model was regressed with `AssertionError: only single axis
transpose supported at this time, decompose in tvm` and fixed it in this
[PR](https://github.com/tenstorrent/tt-forge-fe/pull/2639) by performing
sanity before forge compilation and also noticed generated relay IR
difference between with and without sanity run and found out constant
folding is happening with sanity run and also if the sanity run is
skipped, framework model output remains in float32 rather than bfloat16.

After further analysis found out, the yolov5 model has [Detect
module](https://github.com/ultralytics/yolov5/blob/2540fd4c1c2d9186126a71b3eb681d3a0a11861e/models/yolo.py#L72C1-L72C25)
which has
[grid](https://github.com/ultralytics/yolov5/blob/2540fd4c1c2d9186126a71b3eb681d3a0a11861e/models/yolo.py#L86)
and
[anchor_grid](https://github.com/ultralytics/yolov5/blob/2540fd4c1c2d9186126a71b3eb681d3a0a11861e/models/yolo.py#L87)
tensor initialised in the init function and at first inference, grid and
anchor_grid tensor are computed using make_grid when the grid and input
height and width are not same(i.e [if self.dynamic or
self.grid[i].shape[2:4] !=
x[i].shape[2:4]:](https://github.com/ultralytics/yolov5/blob/2540fd4c1c2d9186126a71b3eb681d3a0a11861e/models/yolo.py#L101))
and at the second inference it will not compute grid and anchor_grid
tensor since the grid is already computed and matches with input height
and width which leads to different graph IR generation so it is not
exactly constant folding.

The framework model output remains in float32 instead of bfloat16 this
problem was also raised in grid and anchor_grid present in the Detect
module. If don't perform sanity run, the model is in bfloat16 and in the
ConvertEmulatedDtypes class we will converts params, buffers, inputs and
tensor peresent in the init function from bfloat16 to float32 before TVM
Relay IR generation but the grid and anchor_grid are list of tensor in
bfloat16 dtype present in the init function which is not converted from
bfloat16 to float32 and and at the same time we do inference at the jit
trace, the grid and anchor are computed using make_grid since the grid
and input height and width are not same which converts the grid and
anchor_grid from bfloat16 to float32 because the grid and anchor_grid
dtypes are determined by anchors buffer which is in float32 and after
TVM graph generation all the params, buffers, inputs and tensors present
in the init function are converted back to bfloat16 from float32 but the
grid and anchors are not converted and it will be in float32 which leads
model output remains in float32 instaed of bfloat16 in forge
verification. To resolve this issues, added support for converting the
list/tuple of tensor present in the init function from bfloat16 to
float32 before TVM graph generated and vice versa in the
ConvertEmulatedDtypes class. Now all the yolov5 model tests runs without
sanity run before forge compilation.



Logs:

[yolov5_before_fix.log](https://github.com/user-attachments/files/21511114/yolov5_before_fix.log)

[yolov5_after_fix.log](https://github.com/user-attachments/files/21511105/yolov5_after_fix.log)