add info about hugepages, new docker link, minor edits for clarity, tested all 3 installs, filed bug for build from source (#2720)
Tested all three methods of building TT-Forge-FE. Build from source had
an error and I filed a ticket with the error output -
https://github.com/tenstorrent/tt-forge-fe/issues/2719. The others work
fine. Work included testing a model with a bug and the correction for
that can be seen here -
https://github.com/tenstorrent/tt-forge/issues/318.

Updates were made to add hugepages configuration info, update the new
docker link for the container, and minor edits for clarity.

### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-forge/issues/317)

### Problem description
All installation steps need to be tested to see if there are new
dependencies or additions that break the installation documentation.

### What's changed
Manually test every method for installation and correct what's broken.
The impact is continuing to have stable, consistent documentation that
yields success on install.

### Checklist
- [ ] New/Existing tests provide coverage for changes