Fix: MaxPool2d padding must be symmetric issue in Googlenet onnx Model (#2777)
### Ticket
[Link to Github
Issue](https://github.com/tenstorrent/tt-forge-fe/issues/2757)

### Problem description
The issue on Maxpool2d which is casued due to the asymmetric padding
case.

### What's changed
The current update includes a revised test script for the ONNX model
export using the pytorch script which result in the correct maxpool2d
usage as the pytorch also does not support the asymmetric case but onnx
does support asymmetric padding case so Instead of loading the model
from a direct link from onnx , it now uses the GoogLeNet PyTorch model
for export and compilation which resolves the issue.
### Checklist
- [x] New/Existing tests provide coverage for changes

### Model Log
Before Fix

[googlenet_log.log](https://github.com/user-attachments/files/21679799/googlenet_log.log)

After Fix

[google_net.log](https://github.com/user-attachments/files/21679796/google_net.log)