Fix resize2d and upsample2d op (#2800)
The PR will fix the following issues in the resize2d and upsample2d ops.

1) In the resize2d eval function, when the operation was an upsample we
were calling the interpolation operation with a scalar factor
(calculated from input shape and target size), while for downsample we
called interpolation with the size argument which doesn't make any sense
so removed the upsample/downsample condition and unified the eval to
always call interpolation with the size argument for both upsampling and
downsampling.

2) In the resize2d shape function, added the integer scalar factor only
supported and resize2d op doesn't support one spatial dimension as
upsample and another as downsample validation checks.

3) In the resize2d post initial decomposition, while decomposing it to
upsample2d op, we are calculating scalar factor from height passing to
upsample2d op which leads to eval and shape mismatch when the height and
width are different scalar factor. So calculated 2 scalar factor from
height and weight passing it as vector of int because the ttir.upsample
op accepts scalar factor as int/vector of int and it also support
different scalar factor. And also update the upsampl2d op eval and shape
function to support both int and vector for scalar factor.

4) The resize2d op has align_corners argument but it was not properly
handled in resize2d eval function and so handled it and at the same time
align_corners argument was not supported in ttnn.upsample2d op so added
the condition to thrown an error while decomposing resize2d to upsampl2d
op.

5) Removed unnecessary conversions between resize method and integer
representations in the resize2d op.

6) Updated the gtest of the resize2d and upsampl2d ops and also updated
the resize2d model ops tests arguments.

7) Added xfail markers for resize2d models ops tests which has
align_corners argument set to True because the ttnn.upsample2d op
doesn't support align_corners argument


### Logs of resize2d:

[test_resize2d.log](https://github.com/user-attachments/files/21776209/test_resize2d.log)


[test_interpolate.log](https://github.com/user-attachments/files/21776207/test_interpolate.log)