Support test-expression input in test workflow (#2923)
### Summary
Add a new workflow input `test-expression` that lets developers run a
subset of pytest tests using a substring expression. This exposes
pytest’s -k matching as an easy-to-use workflow parameter to quickly
target tests by name.

### Motivation
We're enabling training support in the model-analysis pipeline, and in
those scenarios we sometimes need to run either inference or training
test subsets. test-expression makes it possible to select the
appropriate subset