Enhance models analysis to support training (#2955)
1. Currently, training support was added in the model ops tests
generation pipeline. For that, we introduced a pytest parameter called
`training_test`. If training_test is set to True, it will run the model
ops training tests for all the configs present in the
`forge_modules_and_shapes_dtypes_list`. If the inference/training tests
fail with a data mismatch or segmentation fault, we will add
**xfail**/**skip** markers. However, if the inference tests pass for one
config and the training tests fail for the same config, we cannot add
xfail/skip markers because that would also mark the inference tests as
xfail/skip, since the same config list is shared between inference and
training. To resolve this, we introduced two new markers,
`xfail_if_models_ops_training` and `skip_if_models_ops_training`, which
are applied only to training tests. For example, if one config is
failing for training tests alone, we will add
`xfail_if_models_ops_training` for that config. The
`pytest_collection_modifyitems` hook will get the marker, check if
training_test is set to True, and then add the xfail markers for those
configs.

2. Added both **training** and **inference** support in **model analysis
failure updation**. Now it will create a model analysis report which
contains a new column named `test_kind`, which can be either `inference`
or `training`. The model analysis script takes log files to generate the
report and update markers. Since we are currently running the training
and inference model ops tests separately, most of the time, we can't
pass both inference and training test logs. If the inference test logs
are passed to generate the report and update markers, it will check
whether there are any training markers present in the source test file,
extract the information, and update the configs with training markers.
For example, if one config is failing for both training and inference,
we will not add both `xfail` and `xfail_if_models_ops_training` to that
config; we will only add xfail markers. Similarly, cases for inference,
training, and both training and inference are handled.