updates for release, add Python 3.11 (#2971)
### Ticket

https://github.com/tenstorrent/tt-forge/issues/327?issue=tenstorrent%7Ctt-forge%7C481

### Problem description
TT-Forge-FE runs with Python 3.11 now, and instead of needing two
wheels, you now only need one. There is also a new docker container.

### What's changed
Describe the approach used to solve the problem.
The docker container and the wheel were tested on a clean VM. The
instructions (at this time) are up to date and show how to install each
of these.

### Checklist
- [ ] New/Existing tests provide coverage for changes