[build] fix editable installation (#3082)
More recent pip versions are causing issues when installing `forge`
package in editable installations. The new mechansim of path finders
does not work correctly when the package directory is placed in a
directory of the same name, as in our case: `forge/forge`.

More details in the comments.

Added a step in CI to verify that the editable installation is correct
(sanity check).

For some reason, the pre-commit `isort` check is going wild - i am
removing it from pre-commit.

Closes #3081